Modify responsability of calling view
A small modification. Now, discover_view just passes the view to telegramAPI, then telegramAPI passes the messsage to the view. To ensure comomn behavior to future APIs, the base plataform can take more responsability, so Telegram passses the message to BasePlataform. 
Also, when calling the view, the new function also checks if the active pattern has a function to call views (seeing if a method called "call_view" exists). If it exists,  passes the responsability to call view to Pattern. 
This flow will allow the Pattern to make decisios about the flow having access to the return of the view, allowing more complex and flexible interactions.